refactor(quantum_gates): extract measurement into quantum_measure.c
Split the measurement cluster (quantum_measure_multi, quantum_peek_probability,
quantum_measure_all_fast) out of the 1021-line quantum_gates.c into a focused
quantum_measure.c (per soul.md §10). Gate application + verification stay in
quantum_gates.c.

The 3 tiny bit/qubit index helpers (check_qubit_valid, flip_bit, get_bit)
move to a new quantum_bit_utils.h as static-inline, shared by both modules
(DRY - no hand-duplicated one-liners, no cross-TU static dependency).

Verified: libs build clean (the pre-existing M_PI-redefine/format warnings in
bell_test.c/grover.c/matrix_math.c are unrelated to this split); qrng_v3_test
and secure_rng_test pass identically to baseline (15 + 23 PASS). New modules
auto-compiled/linked via the repo's wildcard src/quantum_rng/*.c rule.
